% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage20-package.R
\docType{package}
\name{testpackage20-package}
\alias{testpackage20}
\alias{testpackage20-package}
\title{testpackage20: My Package as a Demo Package}
\description{
What the package does still tbd
}
\author{
\strong{Maintainer}: Shan Lin \email{sl4614a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
